common: FreeBSD compatibility fixes
Use long for level in fs_entry (common/fs.c).
Add FreeBSD stubs for badblock handling.
Close files in cto_dirty test.
Define FALLOCATE in unittest.sh, use in util_badblock test.
Anonymize offset in vmem_valgrind_region memcheck4.log.match.

+++ src/common/fs.h
@@ -59,7 +59,8 @@ struct fs_entry {
 	const char *path;
 	size_t pathlen;
 	/* the depth of the traversal */
-	short level;
+	/* XXX long on FreeBSD. Linux uses short. No harm in it being bigger */
+	long level;
 };
 
 struct fs *fs_new(const char *path);
